Correct path to `inflections.yml` in USAGE.md
The inflections.rb file included in the example wasn't reading the
config/infections.yml file, leading to problems loading the project.

With Rails, inflections.yml is stored in the `config` directory, and the
example in the USAGE.md would look in the top level Rails.root directory.

This change updates the path to the file in a few places in USAGE.md