move ReferenceLister require statement
The require statement for the ReferenceLister class is moved into the classes that
uses the class instead of being required in the packwerk.rb file.

This change is made to make it clearer what dependencies each file has.
It also prevents an issue where the order in the packwerk.rb file needs
to be correct to make the dependencies available. It can have a drawback
of circular dependencies, but that should not be an issue for this
interface.