move file_processor out of the RunContext class into the Cli class
The RunContext class is used to build a NodeProcessor and a FileProcessor. The way the latter one is build is a bit backward because it uses gets initialized in with a RunContext, but used in the CLI class. So I moved the initialization of the FileProcessor out of the RunContext, which made the responsibilities a bit clearer and code easier to follow.