Use new Inflector constructor in run context
`payment_details` in `custom_inflections.yml` is an "uncountable",
which means it should not be inflected when found in a constant association.

Without the fix for this bug, `has_one :payment_details` would not be a
violation because Packwerk is unable to resolve it to a constant and
treats it as a non-violation. In reality, `PaymentDetails` exists as a
constant, which makes it a legitimate violation. See:

https://github.com/Shopify/packwerk/issues/48

In lockstep with this change it was necessary to give the skeleton
application the ability to load its custom inflections, since that’s
what the application validity checker is checking. (This wasn’t a
problem before because Packwerk wasn’t loading the custom inflections
either, but that’s the bug being fixed.)

Co-authored-by: Tom Stuart <tom.stuart@shopify.com>