Use a different custom inflections file in “mismatched inflections” test
We want to be able to use `custom_inflections.yml` to test the
application validation check, since it contains useful inflections that
the app actually uses. When we fix the inflector in a subsequent commit
we’ll need to add `custom_inflections.yml` to its `packwerk.yml`, so we
should use a different set of inflections in this test that’s
intentionally trying to create a mismatch.