add process_file method to RunContext
I'm changing the RunContext class not to leak the file processor object
to the CLI class. This change is reducing how complected the two classes are.

I'm doing this change so that I can pass a NodeProcessorFactory to the
FileProcessor instead of a RunContext instance. If I wouldn't
introducee this change, I would have to make the CLI aware of the
NodeProcessorFactory, which would leak too many implementation details.