Fix CLI init by defining default load path value
`ConfigurationFile#initialize` is expecting an Array value for `load_paths`,
yet when called via the `packwerk init` CLI command, it was being passed as
`nil` and thus throwing a Sorbet type violation.

This restores the prior behaviour of loading the application load paths to
pass to the configuration file generator.

I'm also providing an empty array as a fallback for `load_paths` when loading
a configuration file. This is expected to be an array elsewhere, so this
covers us if the config file doesn't have it defined.