hide "E" on violations when updating deprecations
This may just have been an unintended side effect of the way the code
was previously structured, but when updating deprecations we didn't
display an "E" when we found violations. I liked that behaviour so
this preserves it by adding a single boolean flag to hide errors in
the progress output.