add hash methods for Package
If we add two violations for the same _logical_ package, then they should be added to the same DeprecatedReference instance.

In a single threaded environment, it's always the same package instance, so we don't have these hash misses, but with multiple processes mashalling data, we have a very high liklihood of these misses and thus multiple deprecated reference instances for the same logical package. The consequence of this is incorrect data being written when update-deprecations is run.