add support for specifying default root namespaces
This changes a couple things when running Packwerk against a Rails app.
First of all, instead of getting our load paths via
`config.autoload_paths` and related methods, we use Rails' zeitwerk
autoloaders. Secondly, instead of representing the load paths as a list
of root directory paths, we instead use a hashmap of root directory
paths to module instances which are the root modules for the path.
This allows users to override the default module for any root directory.