Filter custom paths according to include
Custom paths are already filtered by the `exclude` option, but this
commit ensures that the `include` option is respected too. This is
particularly useful for the `packwerk-vscode` extension, which attempts
to run `packwerk check $filename` on save.

For example:

```
include:
  - a/**/*.rb
  - b/**/*.rb
```

I would expect `packwerk check c/some/file.rb` to show the "No files
found or given" message, as it would if `exclude` was set. This commit
ensures that happens.