Allow packwerk to scan for packages inside of engines
This commit modifies packwerk so that it can scan for packages defined
in any engines loaded in a given rails application. This allows
support for more complex repository structures where the rails app is
not at the root of the repository, and where that app depends on
engines defined in sibiling directories.

To support this, I modified `PackageSet` so it also looks at
`Rails.application.railties` to pull a list of engines, and their root
directories, and include those in the paths scanned for packages. I
also dropped the restrictions on load paths for constant resolution in
`ApplicationLoadPaths`. The result is that packwerk can now find
packages defined in these engines, and resolve constants in these
engines.

With this setup, packwerk still only scans files in the actual app and
it's subdirectories for violations, but now it can resolve constants
in the engines, and attach them to packages in the engine.

For testing, I created a new fixture which models this sort of
directory structure, and mocks an app with a rails engine.