Overhaul RESOLVING_VIOLATIONS.md
While I agree that the PDF format wasn't ideal, and mermaid is not a good fit for the complexity of this flow, I think the first markdown version of this is too far from the user experience of the original diagram. This is my attempt at getting a little closer to that.

My approach was to flatten the structure and make it more explicit, as well as add a sprinkle of emoji.