Add associations_exclude to exclude files from association inspection
Other libraries (like Active Model Serializers) use has_one / has_many
macros like Active Record does. We don't need to pay attention to this.
Instead of guessing what a node is based on what we can see in the file,
we are using an exclude glob because it is too hard to know if a class
is an Active Record or not without guessing.